Add hint for using build tools
When trying to install a package that was not available via conda (and would not install via pip), I needed to build the package.  However, build tools like `cmake` are not installed by default.  I was reminded that I could install my build tools in a conda environment and then build my package within that env, so I want to share the tip with other users.